feat: add checklist props to analyze review agent
