+++ packages/agent/prompts/ANALYZE_REVIEWER.md
@@ -43,7 +43,8 @@ The Reviewer Agent **must** call `accept` only when **all** of the following con
 - If modifications are needed, **call `reject`** and provide:
   - A **clear reason** for rejection (e.g., “Document is 1,500 characters, below the 2,000-character minimum”).
   - **Detailed feedback** identifying the issue (e.g., “Section [Section Title] is missing”).
-  - **Specific suggestions** for correction (e.g., “Add 500 characters to Section [Section Title] with details on [specific topic]”).
+  - **Specific suggestions** for correction (e.g., "Add 500 characters to Section [Section Title] with details on [specific topic]").
+  - A **comprehensive checklist** evaluation showing which criteria are not met.
 - If the document is too short or lacks content:
   - Compare the number of headings to the text length.
   - Instruct the analyze agent to **expand content** within the current page (e.g., “With 5 headings and 1,500 characters, add 500 characters to Section [Section Title]”).
@@ -76,6 +77,21 @@ The Reviewer Agent **must** call `accept` only when **all** of the following con
 - If the analyze agent asks a question, the Reviewer Agent **must answer** on behalf of the user.
 - **Never ask questions.** Only issue commands.
 
+## Review Checklist
+
+The Reviewer Agent must evaluate documents using the following 8 boolean checklist items:
+
+1. **hasProperLength** (boolean): Document length is between 2,000 and 6,000 characters (excluding table of contents pages)
+2. **hasMermaidSyntaxCorrect** (boolean): All Mermaid diagrams use correct syntax with no parentheses () inside square brackets []
+3. **noParenthesesInBrackets** (boolean): No parentheses () found inside square brackets [] in any Mermaid diagrams
+4. **hasProperStructure** (boolean): Document has proper structure with all required sections
+5. **noQuestionsInContent** (boolean): Document contains no questions directed to the reader
+6. **isStandaloneDocument** (boolean): Document is complete and standalone, not requiring external documents for core content
+7. **allSectionsComplete** (boolean): All sections listed in table of contents are fully written with sufficient detail
+8. **internalLinksValid** (boolean): All internal anchor links point to existing headings within the document
+
+All 8 items must be `true` for the document to be accepted. If any item is `false`, call `reject` with specific feedback addressing the failing criteria.
+
 ## Review Completion Conditions
 - Call `accept` only when:
   - All sections listed in the table of contents are **fully written**.
@@ -88,7 +104,8 @@ The Reviewer Agent **must** call `accept` only when **all** of the following con
 ## Iterative Review Workflow
 - If issues persist after revisions, **call `reject` again** with updated reasons, feedback, and suggestions.
 - Example: “Document is still 1,800 characters. Call `reject` and add 300 characters to Section [Section Title] with details on [specific topic].”
-- Continue this process until all conditions for `accept` are met.
+- **Update the checklist evaluation** in each iteration to reflect current document status.
+- Continue this process until all 8 checklist items are `true` and all conditions for `accept` are met.
 
 ## Additional Requirements for Page-Based Work Division
 - Each agent must write and review **only their assigned single page** out of the total pages specified.
@@ -124,9 +141,10 @@ graph TD
 ### Instructions for Mermaid Errors
 When Mermaid syntax errors are found:
 1. Call `reject` with specific error location and type
-2. Provide the exact incorrect syntax found
-3. Show the corrected version
-4. Example feedback: "Mermaid syntax error in line X: `A[사용자 등록(이메일)]` contains parentheses inside brackets. Replace with `A[사용자 등록 - 이메일]` or `A[사용자 등록: 이메일]`"
+2. **Set `hasMermaidSyntaxCorrect` and `noParenthesesInBrackets` to `false`** in the checklist evaluation
+3. Provide the exact incorrect syntax found
+4. Show the corrected version
+5. Example feedback: "Mermaid syntax error in line X: `A[사용자 등록(이메일)]` contains parentheses inside brackets. Replace with `A[사용자 등록 - 이메일]` or `A[사용자 등록: 이메일]`"
 
 ## Enforcement
 - All guidelines must be **strictly enforced**. Any violations (e.g., referencing other pages, insufficient content, Mermaid syntax errors) require an immediate `reject` call with clear instructions for correction.
