+++ packages/agent/src/orchestrate/analyze/orchestrateAnalyzeReviewer.ts
@@ -7,16 +7,12 @@ import { AutoBeContext } from "../../context/AutoBeContext";
 import { assertSchemaModel } from "../../context/assertSchemaModel";
 import { enforceToolCall } from "../../utils/enforceToolCall";
 import { AutoBeAnalyzeWriteProps } from "./structures/AutoBeAnalyzeWriteProps";
+import {
+  IAutoBeAnalyzeReviewApplication,
+  IOrchestrateAnalyzeReviewerResult,
+} from "./structures/IAutoBeAnalyzeReviewApplication";
 import { transformAnalyzeReviewerHistories } from "./transformAnalyzeReviewerHistories";
 
-export type IOrchestrateAnalyzeReviewerResult =
-  | {
-      type: "reject";
-      value: string;
-    }
-  | {
-      type: "accept";
-    };
 
 export const orchestrateAnalyzeReviewer = async <
   Model extends ILlmSchema.Model,
@@ -64,34 +60,15 @@ export const orchestrateAnalyzeReviewer = async <
   return fnCalled.value;
 };
 
-interface IAutoBeAnalyzerReviewerSystem {
-  /**
-   * If there is anything that needs to be modified, you can call it, This
-   * function is to reject the document for to try rewriting document with your
-   * advice or suggestion.
-   */
-  reject(input: {
-    /**
-     * The reason why you reject the document and the suggestion for the
-     * modification. You can write the reason in detail.
-     */
-    reason: string;
-  }): "OK" | Promise<"OK">;
-
-  /**
-   * If you decide that you no longer need any reviews, call accept. This is a
-   * function to end document creation and review, and to respond to users.
-   */
-  accept(): "OK" | Promise<"OK">;
-}
 
 function createController<Model extends ILlmSchema.Model>(props: {
   model: Model;
   setResult: (result: IOrchestrateAnalyzeReviewerResult) => void;
 }): IAgenticaController.IClass<Model> {
   assertSchemaModel(props.model);
+
   const application: ILlmApplication<Model> = collection[
-    props.model
+    props.model === "chatgpt" ? "chatgpt" : "claude"
   ] satisfies ILlmApplication<any> as unknown as ILlmApplication<Model>;
   return {
     protocol: "class",
@@ -108,21 +85,22 @@ function createController<Model extends ILlmSchema.Model>(props: {
         props.setResult({
           type: "reject",
           value: input.reason,
+          checklist: input.checklist,
         });
         return "OK" as const;
       },
-    } satisfies IAutoBeAnalyzerReviewerSystem,
+    } satisfies IAutoBeAnalyzeReviewApplication,
   };
 }
 
 const claude = typia.llm.application<
-  IAutoBeAnalyzerReviewerSystem,
+  IAutoBeAnalyzeReviewApplication,
   "claude",
   { reference: true }
 >();
 const collection = {
   chatgpt: typia.llm.application<
-    IAutoBeAnalyzerReviewerSystem,
+    IAutoBeAnalyzeReviewApplication,
     "chatgpt",
     { reference: true }
   >(),
